Issue #58: Refactoring out serialization code so it can be tested separatedly
